refactoring connection into a helper
